Add timeout when connecting with server
I noticed that the thread used to communicate with the watermeter was hanging indefenitely in domoticz. Restarting the thread was not possible, or a disable/enable cycle in the hardware pane neither as it crashed the webserver because it was waiting for the blocked p1 watermeter thread. 
Add a timeout to the timeout call prevents all of this..